pyproject.toml
[project]
name = "product-name-normalizer"
version = "0.1.0"
description = "基于科大讯飞星火大模型的品名标准化工具"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "httpx>=0.27.0",
    "loguru>=0.7.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
